Docs: renamed STEP1.swift
